Add `BinaryPlatforms` module in `Base`
Adds a new `Platform` object that represents a computing platform as a
collection of tags mapping names to values.  Typical tags are things
such as processor `arch`, operating system, which libc is being used on
the platform, etc...  However more specific tags such as what version of
libgfortran the current Julia is linked against or what version of CUDA
is supported by the host are also representable.

Certain tags may require custom comparisons (such as Julia version
number, which should match using only the major and minor version
numbers) and so custom comparators are supported.